requirements.txt
kivy==2.1.0
web3==5.31.3
plyer==2.1.0
cryptography==40.0.0
Pillow==9.4.0
piexif==1.1.3
eth-brownie==1.19.5
eth-abi==2.2.0
opencv-python==4.12.0.88